♻️ 물음표 아이콘 컴포넌트화
