+++ src/components/search/molecules/LabelWithIcon.jsx
@@ -1,0 +1,10 @@
+// LabelWithIcon.jsx
+import styled from 'styled-components';
+
+const LabelWithIcon = styled.label`
+  display: flex;
+  align-items: center;
+  gap: 3px; // 간격 조정
+`;
+
+export default LabelWithIcon;
